requirements.txt
pytest  #Фреймворк для тестирования программного обеспечения на языке Python
selenium #Инструмент автоматизации тестирования веб-приложений
allure-pytest #Адаптер для allure report, который генерирует результаты тестов в формате XML, которые Allure использует для создания отчётов.
webdriver-manager #Библиотека для автоматического управления драйверами браузеров в проектах
modal